requirements.txt
pyodbc
requests
pytz
kafka-python
mysql-connector
aiosfstream